requirements.txt
Django==4.2.5
pytest==7.4.2
pytest-django==4.5.2
pre-commit==3.4.0
pytest-cov==4.1.0
pytest-html==4.0.2
djangorestframework~=3.14.0
pytest-factoryboy==2.5.1
factory-boy==3.3.0
